requirements.txt
asttokens==2.4.1
boto3==1.35.35
botocore==1.35.35
colorama==0.4.6
executing==2.1.0
icecream==2.1.3
jmespath==1.0.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.3
